It kind of works but not really. It needs redone.
In reference to live preview mode inline field rending:

I implemented it using Decorator.mark - big mistake.
The crux: You cannot wrap the mark with a single tag.
Instead of wrapping the whole mark, it wraps each part individually
inside the mark. This means the padding is applied many times when
there is formatting inside the field key or value.

Also certain links don't work? They are parsed correctly.
This has a hard time rendering: [key:: [[file|alias]] ]
and so does this: [key:: [link](https://external.example.com/)]

In addition to that, the '[]' inline fields remain purple as if they
are a link.

In summary: I will have to re-re-do it again in Decorator.replace
			using widgets.

However: This technique will likely work well for implementing source
		 editing mode highlighting. As long as we don't use padding.